boards: stm32h750vb_eft: rename w25q32fv to w25q32fvsig
If spi-bus-width is <4>, since flash_stm32_qspi will recognize
the device as JESD216_MODE_144 mode, while the actual device
only supports JESD216_MODE_114 mode.

When programming in quad mode, the SPI_NOR_CMD_PP_1_4_4 is 0x38,
and 0x38 is in enter QPI mode on w25q32fv, not quad page program,
so the write fails.

The workaround is to make the device work in JESD216_MODE_114 mode
and use SPI_NOR_CMD_PP_1_1_4 as the programming command.

Signed-off-by: canpool <canpool@163.com>